[cmake] Use add_subdirectory() instead of include()
This advances `CMAKE_CURRENT_SOURCE_DIR` in subdirectories as expected. `CMAKE_SOURCE_DIR` keeps pointing to the root of the source-tree.
Using `include()` we treated the whole source-tree as a single directory. Switching to `add_subdirectory()` changes that. Since CMake tracks source file properties by directory, we have to account for it in `set_source_files_properties()`.