Do not try opaque constants when handling infinite loops
We should not break the correctness of the program, even when the
compiler may be able to remove infinite loops with no side-effects.
As constants get reconstructed with a stack-allocated variable, do
not conceal loop-variant constants when possibly in presence of
infinite loops; this otherwise could lead to stack-overflow issues.

Fixes: https://github.com/open-obfuscator/o-mvll/issues/39.