Make Logger thread-safe and write one file per source
The plugin writes logs in many places and they are essential to check
which transformations were applied. The current approach of a single
global logger fails with multithreaded compilers that process one input
per thread.

This patch makes the Logger thread_local, so that we have one instance
per thread and don't need further synchronization. And chooses a unique
name for the log file in two steps:
(1) write to a tmp file with a "-######" hex postfix
(2) rename it to the source file name plug ".omvll" extension